+++ Plugins/AccelByteUe4Sdk/Source/AccelByteUe4Sdk/Public/Api/AccelByteUserApi.h
@@ -1,4 +1,4 @@
-﻿// Copyright (c) 2018-2019 AccelByte Inc. All Rights Reserved.
+﻿// Copyright (c) 2018-2020 AccelByte Inc. All Rights Reserved.
 // This is licensed software from AccelByte Inc, for limitations
 // and restrictions contact your company contract manager.
 
@@ -193,13 +193,13 @@ namespace AccelByte
 			void UnlinkOtherPlatform(const FString& PlatformId, const FVoidHandler& OnSuccess, const FErrorHandler& OnError);
 
 			/**
-			 * @brief This function will search user by their Email Address.
+			 * @brief This function will search user by their Email Address or Display Name.
 			 *
-			 * @param EmailAddress Targeted user's Email Address.
+			 * @param Query Targeted user's Email Address or Display Name.
 			 * @param OnSuccess This will be called when the operation succeeded. The result is FPagedPublicUsersInfo.
 			 * @param OnError This will be called when the operation failed.
 			 */
-			void GetUserByEmailAddress(const FString& EmailAddress, const THandler<FPagedPublicUsersInfo>& OnSuccess, const FErrorHandler& OnError);
+			void SearchUsers(const FString& Query, const THandler<FPagedPublicUsersInfo>& OnSuccess, const FErrorHandler& OnError);
 
 			/**
 			 * @brief This function will search user by userId.