+++ CHANGELOG.md
@@ -17,6 +17,9 @@
     - GetAllUserStatItems
     - GetUserStatItemsByStatCodes
     - GetUserStatItemsByTags
+## Changed
+- GetUserByEmailAddress -> SearchUser
+  - const FString& EmailAddress -> const FString& Query
 
 ## 2.5.0 (2019-11-25)
 ## Fixed